Add metatstudent data files and update README.md
Since metatstudent and loctree3 both need datafiles installed in
order for predictprotein to run properly, they're now downloaded
and installed as part of the Docker image. By doing this, the
image becomes a couple gigbytes large, but saves a lot of work and
time for the end user. The main PredictProtein database is still
necessary to download, since it's quite large and is updated on a
monthly basis.