Include necessary packages instead of relying on external sources
Since third-party repositories may go offline at any time or not
renew their encryption keys, effort has gone in to relying on
external repositories and sources as little as possible in order
to better preserve a reproducable and functioning
`predictprotein-docker` image.

The GitHub "Free" tier limits the size of docker registry images;
therefore, as of this commit and release, we're now self-hosting
the `predictprotein-docker` image at:
https://registry.rostlab.org/repository/rostlab/predictprotein-docker

In addition, this commit introduces a `build.sh` script, which should
be used instead of `docker build`, since `build.sh` will download and
extract a necessary package archive for building the `predictprotein`
docker image. This package archive is hosted on our servers
(rostlab.org), since "free" tier services do not offer the amount
of repository storage necessary.